Fix app crashing on customer delete:
* After customer deletion, redirect to '/athletes' using `Redirect` component, rather than `history.push`
* Add (very dirty) regression test for this (cypress)
* Split 'athlete_form_spec.ts' into 'athlete_management_spec.ts' and 'athlete_self_update.ts' (in 'e2e' package)
