+++ packages/e2e/integration/athlete_management_spec.ts
@@ -1,23 +1,19 @@
 import { Customer } from "@eisbuk/shared";
 import i18n, {
   ActionButton,
-  CustomerNavigationLabel,
   NotificationMessage,
   ValidationMessage,
 } from "@eisbuk/translations";
 
-import { PrivateRoutes, Routes } from "../temp";
+import { PrivateRoutes } from "../temp";
 
 import { customers } from "../__testData__/customers.json";
 import { customers as saulWithExtendedDate } from "../__testData__/saul_with_extended_date.json";
 
 // extract saul from test data .json
 const saul = customers.saul as Customer;
-// Remove the "dial code" from saul's phone
-const saulsDialCode = "IT (+39)";
-const saulsPhone = saul.phone!.substring(3);
 
-describe("add athlete", () => {
+describe("Athlete form", () => {
   beforeEach(() => {
     // Initialize app, create default user,
     // create default organization, sign in as admin
@@ -171,117 +167,36 @@ describe("add athlete", () => {
   });
 });
 
-describe("athlete profile", () => {
+describe("Delete an athlete", () => {
   beforeEach(() => {
-    // Initialize app, create default user,
-    cy.initAdminApp().then((organization) =>
-      cy.updateCustomers(organization, customers as Record<string, Customer>)
-    );
-    cy.signOut();
-
-    cy.visit([Routes.CustomerArea, saul.secretKey].join("/"));
-    cy.clickButton(i18n.t(CustomerNavigationLabel.Profile) as string);
-    // Toggle edit mode
-    cy.clickButton(i18n.t(ActionButton.Edit) as string);
-    // Wait for edit mode to be enabled before making furter assertions
-    cy.get("input").first().should("be.enabled");
+    cy.initAdminApp()
+      .then((organization) =>
+        cy.updateCustomers(organization, customers as Record<string, Customer>)
+      )
+      .then(() => cy.signIn())
+      .then(() => cy.visit(PrivateRoutes.Athletes));
   });
 
-  it("fills and submits athlete profile form", () => {
-    cy.getAttrWith("name", "name").should("have.attr", "value", saul.name);
-    cy.getAttrWith("name", "name").clearAndType(saul.name);
+  it("deletes the customer and redirects back to athletes page", () => {
+    // Open Saul's details
+    cy.contains(saul.name).click();
 
-    cy.getAttrWith("name", "surname").clearAndType(saul.surname);
-    cy.getAttrWith("name", "birthday").clearAndType(saul.birthday || "");
-    cy.getAttrWith("name", "email").clearAndType(saul.email || "");
-    cy.getAttrWith("id", "dialCode").select(saulsDialCode);
-    cy.getAttrWith("name", "phone").clearAndType(saulsPhone);
-    cy.getAttrWith("name", "certificateExpiration").clearAndType(
-      saul.certificateExpiration || ""
-    );
-    cy.getAttrWith("name", "covidCertificateReleaseDate").clearAndType(
-      saul.covidCertificateReleaseDate || ""
-    );
-    cy.getAttrWith("name", "covidCertificateSuspended").click();
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(NotificationMessage.CustomerProfileUpdated) as string);
-  });
+    // Fire the delete action
+    cy.get("button")
+      .contains(i18n.t(ActionButton.DeleteCustomer) as string)
+      .click();
+    // Confirm the prompt
+    /**
+     * @WARNING This is a very dirty hack (depending on the view having exactly 14 buttons),
+     * but cy, for some reason couldn't find the button by its (exact) text: "Delete".
+     */
+    cy.get("button").eq(13).click();
 
-  it("allows customer form submission with minimal fields", () => {
-    cy.getAttrWith("name", "certificateExpiration").clear();
-    cy.getAttrWith("name", "covidCertificateReleaseDate").clear();
-
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(NotificationMessage.CustomerProfileUpdated) as string);
-  });
-
-  it("doesn't allow invalid date input format", () => {
-    cy.getAttrWith("name", "birthday").clearAndType("12 nov 2021");
-
-    cy.getAttrWith("type", "submit").click();
-    // check invalid date message
-    cy.contains(i18n.t(ValidationMessage.InvalidDate) as string);
-  });
-
-  it("doesn't allow invalid phone input format", () => {
-    // test phone number for edge cases
-    cy.getAttrWith("name", "phone").clearAndType("foo 2222 868");
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(ValidationMessage.InvalidPhone) as string);
-
-    cy.getAttrWith("name", "phone").clearAndType("2222 868 foo");
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(ValidationMessage.InvalidPhone) as string);
-
-    // test too long and too short phone numbers
-    cy.getAttrWith("name", "phone").clearAndType("2222 86877777777777");
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(ValidationMessage.InvalidPhone) as string);
-
-    cy.getAttrWith("name", "phone").clearAndType("2222");
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(ValidationMessage.InvalidPhone) as string);
-
-    // make sure phone number length can't be "cheated" with too much whitespace
-    cy.getAttrWith("name", "phone").clearAndType("099   11");
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(ValidationMessage.InvalidPhone) as string);
-  });
-
-  it("replaces different date separators ('.' and '-') with '/'", () => {
-    // dashes
-    cy.getAttrWith("placeholder", "dd/mm/yyyy")
-      .first()
-      .clearAndType("12-12-1990");
-    cy.getAttrWith("value", "12/12/1990");
-
-    // dots
-    cy.getAttrWith("placeholder", "dd/mm/yyyy")
-      .first()
-      .clearAndType("12.12.1990");
-    cy.getAttrWith("value", "12/12/1990");
-  });
-
-  it("handles edge (passable) cases of input", () => {
-    const archer = {
-      // test two names string (should be passable)
-      name: "Sterling Malory",
-      surname: "Archer",
-      // test whitespaces in the phone number
-      // (should be removed in submitting function passable)
-      phone: "99 6622 545",
-      // check insanely long, but passable email
-      email: "sterling.malory.archer@isis.not-gov.us",
-      birthday: saul.birthday,
-    };
-    cy.getAttrWith("name", "name").clearAndType(archer.name);
-    cy.getAttrWith("name", "surname").clearAndType(archer.surname);
-    cy.getAttrWith("name", "phone").clearAndType(archer.phone);
-    cy.getAttrWith("name", "email").clearAndType(archer.email);
-    cy.getAttrWith("name", "birthday").clearAndType(archer.birthday || "");
-
-    // all of the data above should be submitable
-    cy.getAttrWith("type", "submit").click();
-    cy.contains(i18n.t(NotificationMessage.CustomerProfileUpdated) as string);
+    // Should get redirected to the /athletes page and, containing all of the customers except Saul
+    cy.url().should("include", PrivateRoutes.Athletes);
+    Object.values(customers)
+      .filter(({ id }) => id !== saul.id)
+      .forEach(({ name }) => cy.contains(name));
+    cy.contains(saul.name).should("not.exist");
   });
 });
